feat: Implement post-reboot installation step
This is still a work-in-progress. There is currently no mechanism to signal how to set up the target system.

The post-reboot installation step is run once after the first boot via systemd.
The unattended-install.service invokes the post-boot-installation.sh script.
That performs the required installation steps and then deletes itself.

preseed/unattended-mint.seed: Sets up the systemd step that invokes the installer. It also installs openssh-server to (try to) guarantee that the system is remotely accessible even if the post-boot installation steps fail for some reason.

unattended/files/post-boot-installation.sh: The script that performs the required installations.

unattended/files/unattended-install.service: The systemd service description.