feat: Add a script to set everything up from scratch.
unattended/scripts/prepare-aio-area.sh:

This script is very much still a work-in-progress.

Fetch it from github (e.g. using wget) and invoke it to create a work area, clone the repo, download the specified Linux Mint ISO, copy it "under" the repo (i.e. avoid overwriting anything that the repo specifies) and leaves the work area ready to build the unattended install ISO.

There is work to be done to make this more user-friendly and fool proof and more versatile. It could stand a good deal more documentation too. It does, however, basically work.