feat: Get networking active on MSI MAG B550 motherboard.
There's a repo on github that has the RTL8125 driver code available and also a simple script to install it using dkms. Download that while building the AIO ISO and invoke the code as appropriate.

unattended/scripts/prepare-aio-area.sh: Download a repository that contains the RTL8125 driver ...

unattended/.gitignore: ... and make sure that the new repositories directory is ignored by git.

unattended/scripts/handle-r8125.sh: This script loads the RTL8125 driver via dkms and then waits for an IP address to be available (via DHCP).

preseed/unattended-mint.seed: Put the RTL8125 code on the target system and invoke it to make the network available during installation (so that e.g. 'apt update' will work).

unattended/files/post-boot-installation.sh: Invoke the RTL8125 script.