feat: Split EFI and BIOS boot paths.
To make maintenance easier in the future, split up the EFI and BIOS boot pre-seeds. There might be a way to recombine the common parts later on.
Move the partman recipes into separate files as this is easier to read and maintain than the "continued string" format.
The grub (EFI) timeout parameter has not worked for me so far, but set it to "no countdown, wait for user action" anyway.
Set the isolinux (BIOS) timeout to a large value and change the default to be the untouched Linux Mint Live CD.
In passing change "---" in the linux kernel boot line (both isolinux.cfg and grub.cfg) to the "--" (two dashes) that it was always intended to be.