feat: Support the RTL8821CE PCIe Wireless Adapter
The RTL8821CE is built in to some HP laptops. This commit adds support
for it. However, while it seems happy to connect to some networks, it
seems to fail with other networks.

The RTL8821CE driver isn't part of the kernel so has to be installed via
DKMS. That means copying the relevant deb file onto the target machine
and installing the package after the target machine has been rebooted
and is running the final kernel.