feat: Use a config file
Move many of the internal variables that represent configuration
options to a new config.cfg file. The README.md has been updated to
mention the config.cfg file.

Adjust the main scripts to move the script entry point to a main()
function and then invoke that as the last line of the script. This
makes future maintenance easier.