feat: Simplify build
Rearrange the repository so that the build can be simplified. Now it is
necessary only to clone the repo, adjust the config if required and then
run two procedures and a suitable ISO will be produced.

By default all of the output data and downloaded material lives under
bin/. In particular the material that is incorporated into the final ISO
is placed under bin/uai/.