feat: Optionally upgrade to the latest kernel
At the moment the latest Linux Mint installs with a 5.4 kernel. This is
too early to include support for the Vega graphics in an AMD 5600G
processor. However, the latest available and supported kernel (5.15)
does include such support, so this commit provides support for
performing that upgrade.

Note that when first tried, the script tried to use 5.16, which was not
avaiable at the moment for AMD. So rather than let the procedure try for
the latest kernel, I've fixed it at 5.15 for now.

A reasonable enhancement would be to try successively older kernels
until a working one is successfully downloaded. Another simple
enhancement would be to provide a mechanism for specifying the required
kernel.