feat: retry kernel upgrade on failure
The automatic kernel upgrade to the latest version can fail, although
the failure is usually transient (on one occasion I saw the latest
kernel fail to install because it was not available for AMD). To cope
with this scenario, simply drop back to the next most recent kernel and
repeat until a kernel is successfully downlaoded and installed. In fact
the code gives up after either 15 attempts or when the next available
kernel is not newer that then currently installed one.